+++ package.json
@@ -27,7 +27,6 @@
     "react-dom": "18.2.0",
     "react-hook-form": "^7.40.0",
     "react-icons": "^4.7.1",
-    "react-scroll": "^1.8.8",
     "ts-node": "^10.9.1",
     "typescript": "4.9.3"
   }
