+++ src/components/templates/HomeTemplate.tsx
@@ -1,8 +1,8 @@
 import { Box, Button, Flex } from "@chakra-ui/react";
 import { PostBox } from "components/organisms/PostBox";
-import { TimeLine } from "components/organisms/TimeLine";
 import { Layout } from "layouts/Layout";
 import React from "react";
+import { PostList } from "components/organisms/PostList";
 
 export const HomeTemplate: React.FC = () => {
   const returnTop = () => {
@@ -24,18 +24,12 @@ export const HomeTemplate: React.FC = () => {
         borderBottom={"1px solid gray"}
         cursor={"pointer"}
         onClick={returnTop}
+        fontWeight={"bold"}
       >
-        Home
+        HOME
       </Box>
       <PostBox />
-      <TimeLine />
-      <TimeLine />
-      <TimeLine />
-      <TimeLine />
-      <TimeLine />
-      <TimeLine />
-      <TimeLine />
-      <TimeLine />
+      <PostList />
     </Layout>
   );
 };
